move to enum
