+++ src/shared/adapters/fromImportFile.ts
@@ -1,4 +1,4 @@
-import { Coords, HousingFeatures } from '../types/index.js';
+import { Coords, HousingFeature } from '../types/index.js';
 
 export const adaptCoordFromString = (coords: string): Coords => {
   const [latitude, longitude] = coords.split(',');
@@ -7,7 +7,7 @@ export const adaptCoordFromString = (coords: string): Coords => {
 
 export const adaptHousingFeaturesFromString = (
   features: string
-): HousingFeatures => (features.split(',') as HousingFeatures) ?? [];
+): HousingFeature[] => (features.split(',') as HousingFeature[]) ?? [];
 
 export const adaptBooleanFromString = (value: string): boolean =>
   value.startsWith('true');
