+++ src/shared/types/index.ts
@@ -1,9 +1,9 @@
 export * from './comment.interface.js';
 export * from './coords.interface.js';
-export * from './housing-features.type.js';
 export * from './mock-server-data.type.js';
 export * from './rental-offer.interface.js';
-export * from './type-housing.type.js';
 export * from './user.interface.js';
 export * from './component.enum.js';
 export * from './user-type.enum.js';
+export * from './housing-feature.enum.js';
+export * from './type-housing.enum.js';
