+++ src/shared/types/type-housing.enum.ts
@@ -1,0 +1,6 @@
+export enum TypeHousing {
+  Apartment = 'apartment',
+  House = 'house',
+  Room = 'room',
+  Hotel = 'hotel',
+}
