--- src/shared/types/type-housing.type.ts
@@ -1,2 +1,0 @@
-/** Тип жилья. */
-export type TypeHousing = 'apartment' | 'house' | 'room' | 'hotel';
